+++ restaurant-api/src/main/java/com/splitthefeatures/restaurantapi/service/RestaurantService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface RestaurantService {
 
     List<RestaurantDto> findAll();
-    RestaurantDto findById(Integer id);
+    RestaurantDto findById(Long id);
     List<RestaurantDto> findByDescription(String description);
-    List<RestaurantDto> findByCousineId(Integer cousineId);
+    List<RestaurantDto> findByCousineId(Long cousineId);
 }
